feat(latency): add cache warmth estimator and session transition flow
