+++ src/semantic-router/pkg/extproc/session_transition.go
@@ -1,9 +1,6 @@
 package extproc
 
 import (
-	"time"
-
-	"github.com/vllm-project/semantic-router/src/semantic-router/pkg/latency"
 	"github.com/vllm-project/semantic-router/src/semantic-router/pkg/observability/logging"
 )
 
@@ -36,40 +33,4 @@ func populateSessionTransitionFields(ctx *RequestContext) {
 		ctx.TurnIndex = len(ctx.ChatCompletionMessages) / 2
 		// TODO: populate PreviousModel for Chat Completions once per-turn model history is available.
 	}
-}
-
-// maybeEmitTransitionEvent records a ModelTransitionEvent on model change.
-// Must be called after ctx.TTFTSeconds and ctx.CacheWarmthEstimate are set.
-func maybeEmitTransitionEvent(ctx *RequestContext) {
-	if ctx == nil || ctx.SessionID == "" || ctx.RequestModel == "" {
-		return
-	}
-	if ctx.PreviousModel == "" || ctx.PreviousModel == ctx.RequestModel {
-		return
-	}
-
-	previousResponseID := ""
-	if ctx.ResponseAPICtx != nil {
-		previousResponseID = ctx.ResponseAPICtx.PreviousResponseID
-	}
-
-	evt := latency.ModelTransitionEvent{
-		SessionID:           ctx.SessionID,
-		TurnIndex:           ctx.TurnIndex,
-		FromModel:           ctx.PreviousModel,
-		ToModel:             ctx.RequestModel,
-		TTFTMs:              ctx.TTFTSeconds * 1000,
-		CacheWarmthEstimate: ctx.CacheWarmthEstimate,
-		PreviousResponseID:  previousResponseID,
-		Timestamp:           time.Now(),
-	}
-	latency.RecordTransition(evt)
-	logging.ComponentDebugEvent("session", "model_transition", map[string]interface{}{
-		"session_id":            evt.SessionID,
-		"turn_index":            evt.TurnIndex,
-		"from_model":            evt.FromModel,
-		"to_model":              evt.ToModel,
-		"ttft_ms":               evt.TTFTMs,
-		"cache_warmth_estimate": evt.CacheWarmthEstimate,
-	})
 }
